feat: 优化读取长度 & 边读边压

+++ client.go
@@ -65,7 +65,7 @@ type Options struct {
 	// CompressType gzip
 	CompressType string
 	// CompressLimit 大于该值之后才压缩 单位字节
-	CompressLimit int
+	CompressLimit int64
 }
 
 const (
